requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
pyngrok
python-multipart